Removed unwanted type promotions and wrap-arounds.
include/innocent-pear/bbq.h :
- removed unwanted type promotions in `<<'s
- do_inv_opt_rept(...) now clips value of N for case 4

include/innocent-pear/kthxbai.h ,
include/innocent-pear/orly.h ,
include/innocent-pear/yarly.h ,
include/innocent-pear/host/rly.h :
- removed unwanted type promotions in `<<'s and `>>'s

share/innocent-pear/doge-01.cc :
- for ia16-elf, now directly scrambles 0xff as an 8-bit value

test/test-kthxbai-b.cc ,
test/test-kthxbai-b.good :
- new; tests for unwanted type promotions affecting results

GNUmakefile :
- added test-kthxbai-b test case

+++ share/innocent-pear/doge-01.cc
@@ -62,7 +62,7 @@ innocent_pear_DOGE unscramble_01_1()
 	__asm __volatile("outb %0, $0x21"
 	    : /* no outputs */
 	    : "Ral" ((uint_least8_t)
-		kthxbai?<unsigned, allow_for_startup>(0xffu)));
+		kthxbai?<uint_least8_t, allow_for_startup>(0xffu)));
 #elif defined innocent_pear_HAVE_CONST_TCOOFF
 	innocent_pear::rofl?<allow_for_startup, 1u>::
 	    tcflow((int)((0ull + innocent_pear_DOGE_STATE_0 +
